+++ src/components/Buttons/SaveAndEditButton.sc.js
@@ -1,0 +1,15 @@
+export const saveButton = {
+  padding: "0.75rem 1.5rem",
+  fontSize: "1rem",
+  fontFamily: "Montserrat",
+  fontWeight: "500",
+  backgroundColor: "#1565C0",
+  color: "white",
+  border: "none",
+  borderRadius: "1.0625rem",
+  cursor: "pointer",
+  display: "flex",
+  alignItems: "center",
+  gap: "0.5rem",
+  whiteSpace: "nowrap",
+};
