+++ src/reader/TopToolbar.sc.js
@@ -9,22 +9,6 @@ export const middleGroup = {
   gap: "1rem",
 };
 
-export const saveButton = {
-  padding: "0.75rem 1.5rem",
-  fontSize: "1rem",
-  fontFamily: "Montserrat",
-  fontWeight: "500",
-  backgroundColor: "#1565C0",
-  color: "white",
-  border: "none",
-  borderRadius: "1.0625rem",
-  cursor: "pointer",
-  display: "flex",
-  alignItems: "center",
-  gap: "0.5rem",
-  whiteSpace: "nowrap",
-};
-
 export const rightGroup = {
   display: "flex",
   alignItems: "center",
